% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaEncoder.R
\name{betaEncoder}
\alias{betaEncoder}
\title{Likelihood encoding of categorical feautres and binary target variable.}
\usage{
betaEncoder(target, train_group, test_group, prior_strength = 1)
}
\arguments{
\item{target}{target variable (default is \code{NULL})}

\item{group}{group variable}

\item{seed}{random seed (default is \code{sample(1e6, 1)})}
}
\description{
This function uses likelihood encoding to map categorical variables in to 
a numeric value such.  The beta distrubution is used as a prior (the 
conjugate prior of the bernoulli distribution).  The parameters of the 
prior are taken to be \eqn{alpha = n*p} and \eqn{beta = n*(1-p)} where
\eqn{n} is the population size and \eqn{p} is the population mean.  New
values are mapped to a random value from the \eqn{Beta(alpha+a, beta+b)} 
with \eqn{a} being number of successes in the group and \eqn{b} being number
failures.
}
