% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbUtils.R
\name{dt2lgb}
\alias{dt2lgb}
\title{Wrapper to convert a data.table to an lgb.Dataset}
\usage{
dt2lgb(DT, ...)
}
\arguments{
\item{DT}{data.table to convert}

\item{p}{value to pass to \code{overSample} (default is \code{p=0})}
}
\description{
Wrapper to convert a data.table to an lgb.Dataset
}
