% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbUtils.R
\name{dt2xgb}
\alias{dt2xgb}
\title{Wrapper to convert a data.table to an xgb.DMatrix.}
\usage{
dt2xgb(DT, ...)
}
\arguments{
\item{DT}{data.table to convert}

\item{p}{value to pass to \code{overSample} (default is \code{p=0})}
}
\description{
Wrapper to convert a data.table to an xgb.DMatrix.
}
