% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Convenience function for saving data.}
\usage{
saveData(dt, prefix, suffix = NULL, folder = NULL, subfolder = NULL,
  file_ext = "csv", overwrite = TRUE, verbose = TRUE,
  data_path = "./input", ...)
}
\arguments{
\item{prefix}{Either the full name of the file to save (e.g.,
\code{"train_features.csv"} or the prefix of the file to save.  If the full
name is used then the \code{suffix} parameter should be \code{NULL} and the
file extension should be included.}

\item{suffix}{Suffix of the data file to save This parameter could be useful
when interating through features; e.g., in a for loop. (default is \code{NULL})}

\item{folder}{Folder within \code{"./input/"} to save
the file in. (default is \code{NULL})}

\item{subfolder}{Subfolder with in \code{folder} to save the file in.
(default is \code{NULL})}

\item{file_ext}{Determines the file extension of the file to save
\code{saveData()} calls the appropriate function to save the data.
Accepted values are \code{"csv", "rds", "rda", "xgb"}.  If
\code{file_ext = "xgb"} then the file to save is expected to be an
xgb model.  (default is \code{"csv"})}

\item{overwrite}{Whether or not an existing file with the same name should be
overwritten. If \code{overwrite=NA} then the user is asked whether the file
should be overwritten.  (default is \code{TRUE})}

\item{verbose}{Whether or not to print details about the file being saved
(default is \code{TRUE})}

\item{...}{Additional parameters to pass the save function.}
}
\description{
The function saves the desired data in the \code{"./input/*"} directory.
The name of data file will adhere the following naming convention:
\code{prefix} and a \code{suffix} seperated by an underscore.
Use dashes to seperate terms in the \code{prefix} and \code{suffix}.
For example, \code{"train-stage1_tfidf-sublinear.csv"}.
}
\examples{
# save "train_features.csv" in ./Data/clusters/skmeans
saveData("train_features", folder="clusters", subfolder="skmeans")

# save "train_features.csv"
saveData("train", suffix="features")

# save "train_features.rds"
saveData("train", suffix="features", file_ext="rds")

}
\seealso{
\code{\link{loadData}} The complimentary function to \code{saveData}.
}
\keyword{save}
