% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetEncoder.R
\name{targetEncoder}
\alias{targetEncoder}
\title{Mean encoding of categorical features}
\usage{
targetEncoder(target, group)
}
\arguments{
\item{target}{target variable (default is \code{NULL})}

\item{group}{group variable}

\item{seed}{random seed (default is \code{sample(1e6, 1)})}
}
\description{
Mean encoding of categorical features
}
